Resolve labels to relative offsets
Also fix test output which was mismatched for #0 -> $0 and update Core to use
a boxed slice instead of a Vec (no resizing).

It seems likely there will need to be a "resolved" and "unresolved" view of
the core, one for use with load file format + emulation, and one for parse
validation.